+++ src/main/java/com/pax/market/api/sdk/java/api/terminal/dto/TerminalGroupInfoDTO.java
@@ -1,0 +1,49 @@
+/*
+ * *******************************************************************************
+ * COPYRIGHT
+ *               PAX TECHNOLOGY, Inc. PROPRIETARY INFORMATION
+ *   This software is supplied under the terms of a license agreement or
+ *   nondisclosure agreement with PAX  Technology, Inc. and may not be copied
+ *   or disclosed except in accordance with the terms in that agreement.
+ *
+ *      Copyright (C) 2017 PAX Technology, Inc. All rights reserved.
+ * *******************************************************************************
+ */
+package com.pax.market.api.sdk.java.api.terminal.dto;
+
+import java.io.Serializable;
+
+/**
+ * Terminal group info in terminal detail response
+ */
+public class TerminalGroupInfoDTO implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    private Long groupId;
+    private String groupName;
+
+    public Long getGroupId() {
+        return groupId;
+    }
+
+    public void setGroupId(Long groupId) {
+        this.groupId = groupId;
+    }
+
+    public String getGroupName() {
+        return groupName;
+    }
+
+    public void setGroupName(String groupName) {
+        this.groupName = groupName;
+    }
+
+    @Override
+    public String toString() {
+        return "TerminalGroupInfoDTO{" +
+                "groupId=" + groupId +
+                ", groupName='" + groupName + '\'' +
+                '}';
+    }
+}